# api/openapi.yaml
openapi: 3.0.3
info:
  title: WaxDeck API
  version: 0.2.0
  description: |
    First-party REST API for WaxDeck: the self-hosted player, library manager,
    and metadata completer for music, podcasts, and audiobooks.

    Current scope: health, an authentication stub, library browse and search,
    item detail, artwork, lyrics, play-info, playback state (resume), listen
    session ingest, and library scanning. The shapes (keyset pagination, typed
    PIDs, the structured error model, relative media URLs) are load-bearing
    from day one.

    Conventions:
    - All endpoints live under `/api/v1` (this document's `servers` entry).
    - Item identifiers (`pid`) are type-prefixed ULIDs (`tr-` track, `al-` album,
      `ar-` artist, `pc-` podcast, `ep-` episode, `bk-` audiobook, `jb-` job).
    - List endpoints use opaque keyset cursors (`cursor` in, `nextCursor` out),
      never offsets.
    - Errors are always the `Error` schema. `code` is a stable machine-readable
      string; currently defined codes: `invalid-request`, `unauthenticated`,
      `forbidden`, `not-found`, `conflict`, `internal`, `stream-stale` (a
      stream URL no longer matches the file on disk; re-request play-info),
      and `catalog-maintenance` (the catalog is temporarily handed to a
      maintenance operation; retry shortly). New codes may appear; clients
      must treat unknown codes as opaque.
    - Media URLs returned by the API (e.g. `PlayInfo.url`) are relative to the
...
  contact:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Log in and establish a session
      description: >
        Establishes a session and sets the `waxdeck_session` HttpOnly cookie.
        Native clients use the same endpoint and read the bearer token from the
        response body instead of the cookie. Development stub: any non-empty
        credentials are accepted and map to the built-in development user. Real
        account storage, roles, and OIDC come later.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session established.
          headers:
            Set-Cookie:
              description: "`waxdeck_session` HttpOnly SameSite=Lax session cookie."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
  /library/items:
    get:
      operationId: listItems
      summary: Browse library items
      description: >
        Keyset-paginated list of library items, optionally filtered by media
        type. Ordering is stable (title, then pid) so cursors never skip or
        duplicate.
      tags: [library]
      parameters:
        - name: mediaType
          in: query
          description: Restrict results to one media type.
          schema:
            $ref: "#/components/schemas/MediaType"
        - name: cursor
          in: query
          description: >
            Opaque keyset cursor from a previous page's `nextCursor`. Omit for
            the first page.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum items per page.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: One page of items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemPage"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
  /library/browse:
    get:
      operationId: browseList
      summary: Browse a discovery list
      description: >
        Keyset-paginated discovery lists over the whole library: newest,
        recently added, most played, recently played, random, starred, and
        alphabetical. Play-derived lists reflect the calling user's own
        listening state.
      tags: [library]
      parameters:
        - name: list
          in: query
          required: true
          description: Which discovery list to page through.
          schema:
            $ref: "#/components/schemas/DiscoveryList"
        - name: seed
          in: query
          description: >
            Shuffle seed for the `random` list. The same seed pages through
            the same shuffled order, so paging stays stable. When omitted the
            server picks a fresh seed and returns it as the page's `seed`;
            pass that value back with the cursor for later pages. A cursor is
            only valid together with the same `list` and `seed` it was issued
            for.
          schema:
            type: integer
            format: int64
        - name: cursor
          in: query
          description: >
            Opaque keyset cursor from a previous page's `nextCursor`. Omit for
            the first page.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum items per page.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: One page of the list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemPage"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
...
  /library/rescan:
    post:
      operationId: rescanLibrary
      summary: Start a library rescan
      description: >
        Starts an asynchronous scan of every library root and returns the job
        tracking it. Poll the job to observe progress and completion. Scans
        serialize with other catalog jobs; starting one while another runs
        returns the conflict error. This operation becomes admin-only once
        roles land; clients should be prepared for `forbidden`.
      tags: [admin]
      responses:
        "202":
          description: Scan started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
...
  /items/{pid}:
    get:
      operationId: getItem
      summary: Get one item's detail
      description: Full detail for a single library item.
      tags: [library]
      parameters:
        - $ref: "#/components/parameters/Pid"
      responses:
        "200":
          description: The item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
  /items/{pid}/play-info:
    get:
      operationId: getPlayInfo
      summary: Resolve a playable stream for an item
      description: >
        Returns everything a client needs to start playback: a short-TTL,
        origin-relative stream URL (media-token authenticated, so it is playable
        by bare `<audio>` elements, cast devices, and DLNA renderers that cannot
        send headers), the served format, and expiry. Clients re-request
        play-info when a stream URL expires or the server signals
        `stream-stale`.
      tags: [playback]
      parameters:
        - $ref: "#/components/parameters/Pid"
      responses:
        "200":
          description: Stream resolution for the item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayInfo"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
  /items/{pid}/art:
    get:
      operationId: getItemArt
      summary: Get artwork
      description: >
        Artwork as image bytes: the original at full size, or a square-fit
        thumbnail when `size` is given. Besides item PIDs this endpoint also
        accepts album (`al-`) and artist (`ar-`) PIDs, so search hits can
        render artwork directly. Items without any artwork in their fallback
        chain return 404. Responses carry a content-addressed `ETag`;
        revalidate with `If-None-Match` instead of refetching.
      tags: [library]
      parameters:
        - $ref: "#/components/parameters/Pid"
        - name: size
          in: query
          description: >
            Longest-edge bound in pixels for a thumbnail. Omit for the
            original image.
          schema:
            type: integer
            minimum: 16
            maximum: 2048
        - name: If-None-Match
          in: header
          description: Previously returned `ETag`; a match answers 304 with no body.
          schema:
            type: string
      responses:
        "200":
          description: The artwork image.
          headers:
            ETag:
              description: Content-addressed validator for the returned bytes.
              schema:
                type: string
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        "304":
          description: The cached copy is still current.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
  /items/{pid}/lyrics:
    get:
      operationId: getItemLyrics
      summary: Get an item's lyrics
      description: >
        Lyrics for the item: synced lines when available, an unsynced text
        block otherwise. 404 when the item has no lyrics at all.
      tags: [library]
      parameters:
        - $ref: "#/components/parameters/Pid"
      responses:
        "200":
          description: The item's lyrics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lyrics"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/CatalogMaintenance"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: waxdeck_session
      description: HttpOnly SameSite=Lax session cookie (web).
    bearerAuth:
      type: http
      scheme: bearer
      description: Per-device rotating bearer token (native clients).
...
  schemas:
...
    LoginRequest:
      type: object
      description: Credentials for local-account login.
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          description: Account username.
          example: admin
        password:
          type: string
          minLength: 1
          format: password
          description: Account password.
    LoginResponse:
      type: object
      description: Established session plus a bearer token for native clients.
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: >
            Opaque bearer token equivalent to the session cookie. Native
            clients store it in the OS keychain; web clients ignore it and rely
            on the cookie.
...